Слил с пагинатором, контроллер перевёл на критерии

+++ src/main/java/com/example/internship/helper/PageHelper.java
@@ -13,7 +13,7 @@ public enum PageHelper {
      * @param pageSize количество элементов на странице
      * @return количество страниц
      */
-    public static int calculateTotalPages(int total, int pageSize) {
-        return total / pageSize + (total % pageSize > 0 ? 1 : 0);
+    public static int calculateTotalPages(long total, int pageSize) {
+        return (int) total / pageSize + ((int) total % pageSize > 0 ? 1 : 0);
     }
 }
